+++ src/index.ts
@@ -13,6 +13,7 @@ export * from "./functions/generate-random-integer.js";
 export * from "./functions/generate-random-string.js";
 export * from "./functions/heartbeat.js";
 export * from "./functions/is-local-ip-address.js";
+export * from "./functions/is-valid-url.js";
 export * from "./functions/pad-string-with-null.js";
 export * from "./functions/round-number-down-to-nearest.js";
 export * from "./functions/round-number-to-nearest.js";